backend/requirements.txt
fastapi 
uvicorn 
python-multipart 
openai 
python-dotenv 
websockets
numpy
aiofiles
fastapi[all]
